backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Data Processing
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dateutil==2.8.2

# Export
openpyxl==3.1.2
reportlab==4.0.7
pandas>=2.2.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Working Bank Statement Parser from GitHub
git+https://github.com/dalemarana/bankStatementParser.git@main#egg=bank-statement-parser

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

